Extract agent action callbacks from agents-view.tsx into use-agent-actions hook
Move attachToAgent, startAgent, stopAgent, deleteAgent, handleAgentCreated,
and detachAndClearSelection into a dedicated use-agent-actions.ts hook,
reducing agents-view.tsx from 927 to 844 lines.
